refactor: move atlas version functions out of data folder (#781)
